Rework handle and discovery a bit
Handles taking context/returning errors is a bit annoying, they should always
just have a handle. So change the interface to do that. The network access
stuff was a throwback to where we'd lookup keys by their ID, and load them on
demand. That was bad though as it makes for a bit of a DoS vector when you send
fake JWTs with made up IDs. This moves to a periodic, intentional keyset fetch.

Update discovery to manage refreshing itself, either via manual calls or
optionally spawning a background routine.