First pass at interface redesign
Move away from inline closures on HTTP requests, to a handler interface. This
makes it far clearer what actually has to be implemented, and will help us add
an "out of the box" setup experience. We also get to split the handling of
refresh tokens and stuff up.

Storage stops being a magic hidden thing, and becomes more explicit. This gives
implementers a bit more flexibility on how storage is implemented, and allows
use of some of that data e.g for auditing.

We stop the magic one long session, instead having smaller more focused data
items. This minimizes the risk of data at login session time (e.g the ID) being
useful later, and keeps the scope of each stored item to what it's being used
for.